Completed exercises 4.17-4.22

+++ part4/controllers/blogs.js
@@ -1,30 +1,47 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
-blogRouter.get('/blogs', async (request, response) => {
+blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogRouter.post('/blogs', async (request, response) => {
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes ? body.likes : 0,
+    user: user._id,
   })
 
+  // save the blog
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogRouter.delete('/blogs/:id', async (request, response) => {
+blogRouter.delete('/:id', middleware.userExtractor,  async (request, response) => {
+
+  // verify that the user sending the delete request is the user creating the blog
+  const user = request.user
+  const blogToDelete = await Blog.findById(request.params.id)
+  if ( blogToDelete.user.toString() !== user.id.toString() ) {
+    return response.status(401).json({ error: 'you are not authorized to delete this blog' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
-blogRouter.put('/blogs/:id', async (request, response) => {
+blogRouter.put('/:id', async (request, response) => {
   const body = request.body
 
   const blog = {
